requirements.txt
requests
beautifulsoup4
Jinja2
supabase-py
psycopg2-binary